+++ src/Rn.NetCore.Metrics.Rabbit/Extensions/ServiceCollectionExtensions.cs
@@ -1,17 +1,19 @@
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.DependencyInjection.Extensions;
+using Rn.NetCore.Common.Abstractions;
 using Rn.NetCore.Metrics.Outputs;
-using Rn.NetCore.Metrics.Rabbit.Interfaces;
+using Rn.NetCore.Metrics.Rabbit.Factories;
 
 namespace Rn.NetCore.Metrics.Rabbit.Extensions;
 
 public static class ServiceCollectionExtensions
 {
-  public static IServiceCollection AddRnRabbitMQMetrics(this IServiceCollection services)
+  public static IServiceCollection AddRnMetricRabbitMQ(this IServiceCollection services)
   {
+    services.TryAddSingleton<IDateTimeAbstraction, DateTimeAbstraction>();
     services.TryAddSingleton<IRabbitFactory, RabbitFactory>();
     services.TryAddSingleton<IRabbitConnection, RabbitConnection>();
-    services.TryAddSingleton<IMetricOutput, RabbitMetricOutput>();
+    services.AddSingleton<IMetricOutput, RabbitMetricOutput>();
     
     return services;
   }
