+++ src/Rn.NetCore.Metrics.Rabbit/Factories/RabbitFactory.cs
@@ -1,6 +1,6 @@
 using RabbitMQ.Client;
 using Rn.NetCore.Metrics.Rabbit.Config;
-using Rn.NetCore.Metrics.Rabbit.Interfaces;
+using Rn.NetCore.Metrics.Rabbit.Factories;
 
 namespace Rn.NetCore.Metrics.Rabbit;
