+++ src/Rn.NetCore.Metrics.Rabbit/Rn.NetCore.Metrics.Rabbit.csproj
@@ -25,7 +25,7 @@
   <ItemGroup>
     <PackageReference Include="RabbitMQ.Client" Version="6.3.0" />
     <PackageReference Include="Rn.NetCore.Common" Version="6.0.1.103" />
-    <PackageReference Include="Rn.NetCore.Metrics" Version="6.0.1.102" />
+    <PackageReference Include="Rn.NetCore.Metrics" Version="6.0.2.102" />
   </ItemGroup>
 
 </Project>
